feat: добавляет компонент Header
