+++ src/pages/favorites/favorites.tsx
@@ -1,44 +1,20 @@
 import { Helmet } from 'react-helmet-async';
-import { Link } from 'react-router-dom';
+import { useLocation } from 'react-router-dom';
 
 import Footer from '../../components/layout/footer/footer';
 import FavoriteOffers from '../../components/blocks/favorite-offers/favorite-offers';
+import Header from '../../components/layout/header/header';
 
 function Favorites(): JSX.Element {
+  const location = useLocation();
+
   return (
     <div className="page">
       <Helmet>
         <title>6 cities: favorites</title>
       </Helmet>
 
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link" to="/">
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-              </Link>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a className="header__nav-link header__nav-link--profile" href="#">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                    <span className="header__favorite-count">3</span>
-                  </a>
-                </li>
-                <li className="header__nav-item">
-                  <a className="header__nav-link" href="#">
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
+      <Header location={ location }/>
 
       <main className="page__main page__main--favorites">
         <div className="page__favorites-container container">
