+++ src/pages/login/login.tsx
@@ -1,12 +1,15 @@
 import { Helmet } from 'react-helmet-async';
-import { Link, Navigate } from 'react-router-dom';
+import { Navigate, useLocation } from 'react-router-dom';
 import { useRef, FormEvent } from 'react';
 
 import { loginAction } from '../../services/api-actions';
 import { AuthorizationStatus, AppRoute } from '../../components/consts';
 import { useAppDispatch,useAppSelector } from '../../store/useAppDispatch';
+import Header from '../../components/layout/header/header';
 
 function Login(): JSX.Element {
+  const location = useLocation();
+
   const loginRef = useRef<HTMLInputElement | null>(null);
   const passwordRef = useRef<HTMLInputElement | null>(null);
 
@@ -34,17 +37,7 @@ function Login(): JSX.Element {
         <title>6 cities: authorization</title>
       </Helmet>
 
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link" to={ AppRoute.Main }>
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-              </Link>
-            </div>
-          </div>
-        </div>
-      </header>
+      <Header location={ location }/>
 
       <main className="page__main page__main--login">
         <div className="page__login-container container">
