+++ src/pages/not-found/not-found.tsx
@@ -1,26 +1,19 @@
 import { Helmet } from 'react-helmet-async';
-import { Link } from 'react-router-dom';
+import { Link, useLocation } from 'react-router-dom';
 
 import { AppRoute } from '../../components/consts';
+import Header from '../../components/layout/header/header';
 
 function NotFound(): JSX.Element {
+  const location = useLocation();
+
   return(
     <div className="page page--gray page--login">
       <Helmet>
         <title>6 cities: page is not found</title>
       </Helmet>
 
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link" to={ AppRoute.Main }>
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-              </Link>
-            </div>
-          </div>
-        </div>
-      </header>
+      <Header location={ location }/>
 
       <main className="page__main">
         <div className="container">
